+++ frontend/js/util/utils.ts
@@ -391,7 +391,7 @@ export function downloadAsJSON(object: unknown, filename: string) {
   const blob = new Blob([serialized], { type: "application/json" });
   const url = URL.createObjectURL(blob);
   const link = document.createElement("a");
-  
+
   link.href = url;
   link.download = filename;
   document.body.appendChild(link);